look into cleaning up game object manager
