+++ src/gui/Gui.c
@@ -38,7 +38,7 @@ void DebugMenu(GameObjectManager* gameObjectManager)
 
 	for (size_t i = 0; i < gameObjectManager->lastIndex; i++)
 	{
-		if (gameObjectManager->gameObjects[i]->name != NULL)
+		if (gameObjectManager->gameObjects[i] != NULL && gameObjectManager->gameObjects[i]->name != NULL)
 		{
 			GameObjectPanel(gameObjectManager->gameObjects[i]);
 		}
@@ -62,6 +62,7 @@ void GuiRender()
 
 void GameObjectPanel(GameObject* gameObject)
 {
+	if (gameObject == NULL) return;
 	if (igCollapsingHeader_TreeNodeFlags(gameObject->name, ImGuiTreeNodeFlags_CollapsingHeader))
 	{
 		igPushStyleColor_Vec4(ImGuiCol_Button, (ImVec4) { 0.45f, 0.0f, 0.0f, 1.0f });
@@ -70,6 +71,7 @@ void GameObjectPanel(GameObject* gameObject)
 		if (igButton("Delete", (ImVec2) { 60.0f, 20.0f }))
 		{
 			GameObjectManagerRemove(&gameObjectManager, gameObject->id);
+			return;
 		}
 		igPopStyleColor(3);
 		igCheckbox("Debug", &gameObject->debug);
